requirements.txt
bokeh
boto3
python-dotenv